+++ projects/ngx-libs/src/lib/ngx-perfect-scrollbar/index.ts
@@ -1,0 +1,8 @@
+import { from } from 'rxjs';
+
+export { NgxPerfectScrollbarModule } from './ngx-perfect-scrollbar.module';
+export { NgxPerfectScrollbarComponent } from './ngx-perfect-scrollbar.component';
+export { NgxPerfectScrollbarService } from './ngx-perfect-scrollbar.service';
+export {
+    SCROLL_DIRECTION
+} from './classes/index';