+++ projects/ngx-libs/src/lib/ngx-perfect-scrollbar/ngx-perfect-scrollbar.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { NgxPerfectScrollbarComponent } from './ngx-perfect-scrollbar.component';
+
+@NgModule({
+  imports: [
+    CommonModule
+  ],
+  declarations: [NgxPerfectScrollbarComponent],
+  exports: [NgxPerfectScrollbarComponent]
+})
+export class NgxPerfectScrollbarModule { }