+++ projects/ngx-libs/src/public_api.ts
@@ -8,5 +8,14 @@ export {
     NgxPrismjsDirective,
     NgxPrismjsService
 } from './lib/ngx-prismjs/index';
-export { NgxShadeModule, NgxShadeComponent } from './lib/ngx-shade/index';
+export {
+    NgxShadeModule,
+    NgxShadeComponent
+} from './lib/ngx-shade/index';
+export {
+    NgxPerfectScrollbarModule,
+    NgxPerfectScrollbarComponent,
+    NgxPerfectScrollbarService,
+    SCROLL_DIRECTION,
+} from './lib/ngx-perfect-scrollbar/index';