+++ src/app/modules/docs/docs.module.ts
@@ -3,11 +3,13 @@ import { CommonModule } from '@angular/common';
 import { DocsComponent } from './docs.component';
 import { DocsRoutesModule } from './docs.routing';
 import { DocsService } from './docs.service';
+import { NgxLibsModule } from 'ngx-libs';
 
 @NgModule({
   imports: [
     CommonModule,
     DocsRoutesModule,
+    NgxLibsModule
   ],
   declarations: [DocsComponent],
   providers: [